Mute unused error instead of overriding process()
Instead of overriding the main process() method completley, we can just perform some additional checks and then use the rulset.xml to mute the ShortNoCaptial error that we don't want. Instead, we do our own short comment checks and use different codes so they are not confused.

This sniff isn't 2.x ready yet, but hopefully this will make it easier to become ready.